nodes.get_repositories_metering_info
