+++ specification/nodes/get_repositories_metering_info/GetRepositoriesMeteringInfoResponse.ts
@@ -1,0 +1,41 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { NodesResponseBase } from '@nodes/_types/NodesResponseBase'
+import { RepositoryMeteringInformation } from '@nodes/_types/RepositoryMeteringInformation'
+import { Dictionary } from '@spec_utils/Dictionary'
+import { Name } from '@_types/common'
+import { NodeStatistics } from '@_types/Node'
+
+export class Response extends NodesResponseBase {
+  body: {
+    /**
+     * Contains statistics about the number of nodes selected by the request.
+     */
+    _nodes: NodeStatistics
+    /**
+     * Name of the cluster. Based on the [Cluster name setting](https://www.elastic.co/guide/en/elasticsearch/reference/current/important-settings.html#cluster-name).
+     */
+    cluster_name: Name
+    /**
+     * Contains repositories metering information for the nodes selected by the request.
+     */
+    nodes: Dictionary<string, RepositoryMeteringInformation>
+  }
+}
